Create upload worker
