+++ zboxcore/sdk/multi_operation_worker.go
@@ -3,13 +3,13 @@ package sdk
 import (
 	"context"
 	"fmt"
-	"io"
+	// "io"
 	"sync"
 	"time"
 
-	// "github.com/0chain/errors"
 	thrown "github.com/0chain/errors"
-	"github.com/0chain/gosdk/constants"
+	// "github.com/0chain/gosdk/constants"
+
 	"github.com/0chain/gosdk/core/util"
 	"github.com/0chain/gosdk/zboxcore/allocationchange"
 	"github.com/0chain/gosdk/zboxcore/blockchain"
@@ -55,12 +55,12 @@ func (mo *MultiOperation) Process() error {
 	ctx, ctxCncl := context.WithCancel(context.Background())
 	defer ctxCncl()
 	errs := make(chan error, 1)
-
+	uid := util.GetNewUUID()
 	for idx, op := range mo.operations {
 		wg.Add(1)
 		allocDetails := AllocationDetails{allocationObj: mo.allocationObj, allocationID: mo.allocationID, allocationTx: mo.allocationTx}
 		// Here make it goroutine
-		go func(op Operationer, idx int) {
+		func(op Operationer, idx int) {
 			defer wg.Done()
 
 			// Check for other goroutines signal
@@ -83,7 +83,7 @@ func (mo *MultiOperation) Process() error {
 
 				return 
 			}
-			uid := util.GetNewUUID()
+			// Temporary comment for debugging
 			changes := op.buildChange(refs, uid)
 
 			// change := op.buildChange(refs, uid)
@@ -145,6 +145,7 @@ func (mo *MultiOperation) Process() error {
 		l.Logger.Info("Commit request sending to blobber ", commitReq.blobber.Baseurl);
 		// Here this should be goroutine
 		AddCommitRequest(commitReq)
+		// time.Sleep( 20 * time.Second)
 		cntr++;
 	}
 	wg.Wait()
@@ -171,180 +172,8 @@ func (mo *MultiOperation) Process() error {
 	return nil;
 
 }
-type OperationRequest struct {
-	OperationType string
-	LocalPath string
-	RemotePath string
-	DestName string  // Required only for rename operation
-	DestPath string // Required for copy operation
-
-	// Required for uploads
-	Workdir string
-	FileMeta FileMeta 
-	FileReader io.Reader 
-	Opts []ChunkedUploadOption
-
-}
-
-func CreateMultiOperation(allocObj *Allocation, operations []OperationRequest) (*MultiOperation, error) {
-	var multiOperation MultiOperation;
-	multiOperation.allocationObj = allocObj;
-	multiOperation.allocationID = allocObj.ID
-	multiOperation.allocationTx = allocObj.Tx;
-	multiOperation.blobbers = allocObj.Blobbers
-	multiOperation.operationMask = zboxutil.NewUint128(1).Lsh(uint64(len(allocObj.Blobbers))).Sub64(1)
-	multiOperation.maskMU = &sync.Mutex{}
-	multiOperation.ctx, multiOperation.ctxCncl = context.WithCancel(allocObj.ctx)
-	multiOperation.Consensus = Consensus{
-		consensusThresh: allocObj.consensusThreshold,
-		fullconsensus:   allocObj.fullconsensus,
-	}
-	multiOperation.connectionID = zboxutil.NewConnectionId()
-
-	for _, operation := range(operations) {
-		switch operation.OperationType {
-		case constants.FileOperationRename: 
-			destName := operation.DestName
-			remotePath := operation.RemotePath
-			if destName == "" {
-				return nil, thrown.New("missing parameter", "Rename operation must have destName")
-			}
-			if remotePath == "" {
-				return nil, thrown.New("missing parameter", "Rename operation must have remotePath")
-			}
-			renameOp := &RenameOperation{newName: destName, remotefilepath: remotePath}
-			renameOp.renameMask = multiOperation.operationMask
-			renameOp.maskMU = multiOperation.maskMU
-			renameOp.consensus.consensusThresh = multiOperation.consensusThresh
-			renameOp.consensus.fullconsensus = multiOperation.fullconsensus
-			renameOp.ctx, renameOp.ctxCncl =  context.WithCancel(allocObj.ctx)
-			
-			// multiOperation = append(multiOperation, renameOp);
-			multiOperation.operations = append(multiOperation.operations, renameOp)
-
-		// case constants.FileOperationCopy: 
-		default: 
-			return nil, thrown.New("invalid_operation", "Operation is not valid");
-
-		}
-	}
-	return &multiOperation, nil;
-}
 
 
-
-type UploadOperation struct {
-	workdir string
-	fileMeta FileMeta
-	fileReader io.Reader
-	opts []ChunkedUploadOption
-	refs []fileref.FileRef
-	isUpdate bool
-	
-
-}
-func (uo *UploadOperation) Process(allocDetails AllocationDetails,connectionID string, blobbers []*blockchain.StorageNode) ([]fileref.RefEntity, error){
-	cu, err := CreateChunkedUpload(uo.workdir, allocDetails.allocationObj, uo.fileMeta, uo.fileReader, uo.isUpdate, false, connectionID, uo.opts...);
-	if err != nil {
-		return nil, err
-	}
-	err = cu.process()
-	if err != nil {
-		return nil, err
-	}
-	var pos uint64
-	numList := len(cu.blobbers)
-	objectTreeRefsEntity := make([]fileref.RefEntity, numList)
-	uo.refs = make([]fileref.FileRef, numList)
-	for i := cu.uploadMask; !i.Equals64(0); i = i.And(zboxutil.NewUint128(1).Lsh(pos).Not()) {
-		pos = uint64(i.TrailingZeros())
-		objectTreeRefsEntity[pos] = cu.blobbers[pos].fileRef;
-		uo.refs[pos] = *cu.blobbers[pos].fileRef
-		uo.refs[pos].NumBlocks =  int64(cu.progress.ChunkIndex + 1)
-		uo.refs[pos].ChunkSize = cu.chunkSize
-	}
-
-	l.Logger.Info("Completed the upload")
-	return objectTreeRefsEntity, nil
-}
-
-func (uo *UploadOperation) buildChange(dummyRefs []fileref.RefEntity, uid uuid.UUID) []allocationchange.AllocationChange {
-	changes := make([]allocationchange.AllocationChange, len(uo.refs))
-	for idx, ref := range uo.refs {
-		
-		if uo.isUpdate {
-			change := &allocationchange.UpdateFileChange{}
-			change.NewFile = &ref
-			change.NumBlocks = ref.NumBlocks
-			change.Operation = constants.FileOperationUpdate
-			change.Size = ref.Size
-			changes[idx] = change
-			continue;
-		}
-		newChange := &allocationchange.NewFileChange{}
-		newChange.File = &ref
-		newChange.NumBlocks = ref.NumBlocks
-		newChange.Operation = constants.FileOperationInsert
-		newChange.Size = ref.Size
-		newChange.Uuid = uid
-		changes[idx] = newChange
-	}
-	
-	return changes
-
-}
-
-func (uo *UploadOperation) build(workdir string, fileMeta FileMeta, fileReader io.Reader, isUpdate bool, opts ...ChunkedUploadOption) {
-	uo.workdir = workdir
-	uo.fileMeta = fileMeta
-	uo.fileReader = fileReader
-	uo.opts = opts
-	uo.isUpdate = isUpdate
-}
-
-func (uo *UploadOperation) Verify(allocationObj *Allocation) error {
-	if allocationObj == nil {
-		return thrown.Throw(constants.ErrInvalidParameter, "allocationObj")
-	}
-
-	if !uo.isUpdate && !allocationObj.CanUpload() || uo.isUpdate && !allocationObj.CanUpdate() {
-		return  thrown.Throw(constants.ErrFileOptionNotPermitted, "file_option_not_permitted ")
-	}
-
-	err := ValidateRemoteFileName(uo.fileMeta.RemoteName)
-	if err != nil {
-		return err
-	}
-	spaceLeft := allocationObj.Size
-	if allocationObj.Stats != nil {
-		spaceLeft -= allocationObj.Stats.UsedSize
-	}
-
-	if uo.isUpdate {
-		f, err := allocationObj.GetFileMeta(uo.fileMeta.RemotePath)
-		if err != nil {
-			return err
-		}
-		spaceLeft += f.ActualFileSize
-	}
-	if uo.fileMeta.ActualSize > spaceLeft {
-		return ErrNoEnoughSpaceLeftInAllocation
-	}
-	if uo.isUpdate {
-		otr, err := allocationObj.GetRefs(uo.fileMeta.RemotePath, "", "", "", fileref.FILE, "regular", 0, 1)
-		if err != nil {
-			l.Logger.Error(err)
-			return  thrown.New("chunk_upload", err.Error())
-		}
-		if len(otr.Refs) != 1 {
-			return thrown.New("chunk_upload", fmt.Sprintf("Expected refs 1, got %d", len(otr.Refs)))
-		}
-	}
-	return nil
-
-
-	
-}
